+++ app/components/Section/index.js
@@ -24,6 +24,10 @@ export default class Section extends Component {
     examples: PropTypes.array,
     // React element class used for rendering additional examples
     exampleComponent: PropTypes.func,
+    // Examples config. TODO consider moving all example config in here.
+    defaultTabOpts: PropTypes.shape({
+      firstTabName: PropTypes.string
+    }),
     // Reference to the example component
     _self: PropTypes.object,
     // Id number of the section; used for caching/highlighting purposes
@@ -32,7 +36,8 @@ export default class Section extends Component {
 
   static defaultProps () {
     return {
-      examples: []
+      examples: [],
+      defaultTabOpts: {}
     }
   }
 
@@ -127,8 +132,9 @@ export default class Section extends Component {
 
     // Renders the base example using the styleguide block
     if (this.props.children) {
+      const firstTabName = this.props.defaultTabOpts && this.props.defaultTabOpts.firstTabName || 'Example'
       examples.push(
-        <Tabs.Panel key={'tab-panel-' + exampleId} title={'Example'}>
+        <Tabs.Panel key={'tab-panel-' + exampleId} title={firstTabName}>
           <div className='sg-section-example-body'>
             {this.props.children}
           </div>